refac:검색에 사용되는 변수명 변경
